changes:
CalendarApp-Extra methods for Controlling the app

Day - added a method that checks if an event is "add-able" to a day

Month and NonRepeatingEvents...minor additions like toString() methods
and so on for the sake of testing as we develop.

DBops - added a method that fetches a TreeSet of all the events in the
database

HomeController - major additons...automatic event adding is almost fully
done, check automaticEventAdding in HomeController for more details.

Very minor change made to EventDayIntegrationTesting to the test method
that checks event overlapping. All fixed.